fix: OpenAI Completions API and Fireworks
We were passing a dict into the compat mixin for OpenAI Completions
when using Llama models with Fireworks, and that was breaking some
strong typing code that was added in openai_compat.py. We shouldn't
have been converting these params to a dict in that case anyway, so
this adjusts things to pass the params in as their actual original
types when calling the OpenAIChatCompletionToLlamaStackMixin.

Signed-off-by: Ben Browning <bbrownin@redhat.com>