fix: Together provider shutdown and default to non-streaming
The together inference provider was throwing a stack trace every time
it shut down, as it was trying to call a non-existent `close` method
on the AsyncTogether client. While fixing that, I also adjusted its
shutdown logic to close the OpenAI client if we've created one of
those, as that client does have a `close` method.

In testing that, I also realized we were defaulting to treating all
requests as streaming requests instead of defaulting to
non-streaming. So, this flips that default to non-streaming to match
how the other providers work.

I tested this by ensuring the together inference provider no longer
spits out a long stack trace when shutting it down and by running the
OpenAI API chat completion  verification suite to ensure the change in
default streaming logic didn't mess anything else up.

Signed-off-by: Ben Browning <bbrownin@redhat.com>