feat: refactor external providers dir
currently the "default" dir for external providers is `/etc/llama-stack/providers.d`

This dir is not used anywhere nor created.

Switch to a more friendly `~/.llama/providers.d/`

This allows external providers to actually create this dir and/or populate it upon installation, `pip` cannot create directories in `etc`.

If a user does not specify a dir, default to this one

see https://github.com/containers/ramalama-stack/issues/36

`llama stack build` and `llama stack run` needed to be modified to work with this change and with external providers dir in general.

`llama stack run --image-type container --image-name foobar` should _not_ require a `--config`. This is because the config is copied in during the build and accounts for the external providers dir.

forcing a run yaml at runtime breaks external providers because the host-system path to the external providers is used in the container which is wrong

Signed-off-by: Charlie Doern <cdoern@redhat.com>